requirements.txt
coir-retrieval
google-generativeai
sentence-transformers
torch
numpy
matplotlib
tqdm